+++ lib/types.ts
@@ -288,3 +288,5 @@ export const ONE_DAY = ONE_HOUR * 24;
 export const ONE_WEEK = ONE_DAY * 7;
 export const ONE_MONTH = ONE_DAY * 30;
 export const ONE_YEAR = ONE_DAY * 365;
+
+export type CalcOpt = { fractionDigits: number };