+++ lib/utils/normalize-number.util.ts
@@ -1,6 +1,8 @@
+import IsNaN from "./is-nan.util";
+
 export const Float = (value: number): number => {
-    if(typeof value === 'string' && !isNaN(value)) return parseFloat(value);
-    if(typeof value === 'number') return parseFloat(value.toFixed(5));
+    if(typeof value === 'string' && !IsNaN(value)) return parseFloat(value);
+    if(typeof value === 'number') return value;
     return 0;
 }