+++ lib/utils/sum-number.util.ts
@@ -1,24 +1,27 @@
+import EnsureNumber from "./ensure-number";
+import IsNaN from "./is-nan.util";
 import Float from "./normalize-number.util";
 import ToDecimal from "./to-decimal-number.util";
 import ToLong from "./to-long-number.util";
+import ToPrecision from "./to-precision.util";
 
-export const Sum = (valueA: number, valueB: number): number => {
+export const Sum = (valueA: number, valueB: number, precision = 5): number => {
     const isValueAnumber = typeof valueA === 'number';
     const isValueBnumber = typeof valueB === 'number';
     const isBothNumber = isValueAnumber && isValueBnumber;
 
     if(!isBothNumber){
-        const isNaNValueA = isNaN(valueA);
-        const isNaNValueB = isNaN(valueB);
+        const isNaNValueA = IsNaN(valueA);
+        const isNaNValueB = IsNaN(valueB);
         const isBothNaN = isNaNValueA && isNaNValueB;
         if(isBothNaN) return 0;
         if(isNaNValueA) return Float(valueB);
         if(isNaNValueB) return Float(valueA);
-        if(typeof valueA === 'string' && typeof valueB === 'string') return Float(ToDecimal(ToLong(Float(valueA)) + ToLong(Float(valueB))));
-        if(typeof valueA === 'string' && typeof valueB === 'number') return Float(ToDecimal(ToLong(Float(valueA)) + ToLong(valueB)));
-        if(typeof valueB === 'string' && typeof valueA === 'number') return Float(ToDecimal(ToLong(valueA) + ToLong(Float(valueB))));
+        const result = ToPrecision(Float(ToDecimal(ToLong(Float(valueA)) + ToLong(Float(valueB)))), precision);
+        return EnsureNumber(result);
     }
-    return Float(ToDecimal(ToLong(valueA) + ToLong(valueB)));
+    const result = ToPrecision(Float(ToDecimal(ToLong(valueA) + ToLong(valueB))), precision);
+    return EnsureNumber(result);
 }
 
 export default Sum;